Add symmetry information to forte (#194)
* Switch point group information to forte. Remove deep header dependencies

* Format

* Redefine short test cases and add medium length type

* Fixed issue with wrong key

* Remove CharacterTable from a good portion of forte

* Remove uses of CharacterTable

* Fix compilation error

* Remove unncessary include files

* Add header

* Small edit

* Add missing #include <chrono>

* Add another missing file

* Remove warnings

* Fix pytests

* Add missing variable

+++ src/helpers/iterative_solvers.cc
@@ -388,8 +388,8 @@ void DavidsonLiuSolver::collapse_vectors() {
 std::pair<bool, bool> DavidsonLiuSolver::check_convergence() {
     compute_residual_norm();
     // check convergence on all roots
-    int num_converged_energy = 0;
-    int num_converged_residual = 0;
+    size_t num_converged_energy = 0;
+    size_t num_converged_residual = 0;
     converged_ = 0;
     if (print_level_ > 1) {
         outfile->Printf("\n  Root      Eigenvalue        Delta   Converged?\n");
